moved test utils method
